Move OpenMensa model to root of project
