+++ parsers/aachen/__init__.py
@@ -1,3 +1,2 @@
 from . import model
-from . import openmensa_model
 from .aachen import parser
